bulk load can set null string (postgres)
During a bulk load, by default postgres interprets ',\N,'
as a null, but ',,' as an empty string.  This is different
from CSV flat files, which create ',,' for nulls.

So, if you are loading CSV files, but aren't setting the
other options which put postgres COPY into a CSV mode,
you need to be able to specify '' as the string to interpret
as a null (the null string).

This patch allows setting the null string.

The bulk load helper still doesn't implement the full
postgres COPY interface.  If you want to attempt that,
see http://www.postgresql.org/docs/8.2/static/sql-copy.html.